// src/services/models/person/dto/personRequestDto.ts
export type PersonRequestDto = {
    nombre: string;
    fechaNacimiento: string;
    colorOjos: string;
    genero: string;
    colorCabello: string;
    estatura: string;
    peso: string;
    colorPiel: string;
    planetaNatal: string;
    peliculas: string[];
    especies: string[];
    navesEstelares: string[];
    vehiculos: string[];
    url: string;
}