// src/repositories/models/person.ts
export type Person = {
    id: string,
    creationDate: string,
    name: string;
    birthDate: string;
    eyeColor: string;
    gender: string;
    hairColor: string;
    height: string;
    mass: string;
    skinColor: string;
    homeworld: string;
    films: string[];
    species: string[];
    starships: string[];
    vehicles: string[];
    url: string;
}